backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.34
asyncpg==0.29.0
pydantic-settings==2.3.4
httpx==0.27.2
ortools==9.11.4210
python-dotenv==1.0.1
